[gir2swift] Change project structure
This commit splits long files to smaller ones, introduces subdirectories
and removes some superfluous extensions
