+++ Sources/libgir2swift/emmiting/gtk2swiftdoc.swift
