+++ Sources/libgir2swift/models/Gir+Enums.swift
