+++ Sources/libgir2swift/models/Gir+KnowTypeSets.swift
