+++ Sources/libgir2swift/models/Gir+KnownTypes.swift
