+++ Sources/libgir2swift/models/TypeConversion.swift
