+++ Sources/libgir2swift/models/girtype+xml.swift
