+++ Sources/libgir2swift/utilities/Collection+Utilities.swift
