+++ Sources/libgir2swift/utilities/FileLen.swift
