+++ Sources/libgir2swift/utilities/String+ContentsOfFile.swift
