+++ Sources/libgir2swift/utilities/basename.swift
