+++ Sources/libgir2swift/utilities/cstring.swift
