+++ Sources/libgir2swift/utilities/getopt.swift
