+++ Sources/libgir2swift/utilities/mmap.swift
